fix(timetable): use floats to represent attendance and round them for the UI

+++ app/src/main/java/com/example/vitea/models/timetable/Lecture.kt
@@ -19,6 +19,6 @@ data class Lecture(
     val name: String,
     val course: String,
     val outTime: String,
-    val attendance: Int,
+    val attendance: Float,
     val getAttendanceDetails: List<GetAttendanceDetails>
 )
